Add pvc cache, To #40605168
